pyproject.toml
[tool.poetry]
name = "qubity"
version = "0.1.0"
description = ""
authors = ["Ruslan Sirazhetdinov <ruslansir@icloud.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-binance = "^1.0.19"
pydantic = "^2.9.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
